+++ tests/integration/cases/record_epoch_datetime_i32_overflow/CSharp@v10.cs
@@ -1,0 +1,6 @@
+using System;
+using System.Collections.Generic;
+var my_data = new Dictionary<string, DateTime> {
+    ["within_i32"] = new DateTime(2024, 1, 15, 12, 0, 0),
+    ["beyond_i32"] = new DateTime(2099, 6, 15, 8, 30, 0)
+};