+++ tests/integration/cases/record_epoch_datetime_i32_overflow/Go_combined@v1_18.go
@@ -1,0 +1,14 @@
+package main
+import "time"
+
+func main() {
+my_data := map[string]time.Time{
+	"within_i32": time.Date(2024, time.January, 15, 12, 0, 0, 0, time.UTC),
+	"beyond_i32": time.Date(2099, time.June, 15, 8, 30, 0, 0, time.UTC),
+}
+my_data = map[string]time.Time{
+	"within_i32": time.Date(2024, time.January, 15, 12, 0, 0, 0, time.UTC),
+	"beyond_i32": time.Date(2099, time.June, 15, 8, 30, 0, 0, time.UTC),
+}
+_ = my_data
+}