+++ tests/integration/cases/record_epoch_datetime_i32_overflow/Go_record_epoch_i32_overflow@v1_18.go
@@ -1,0 +1,13 @@
+package main
+type Record0 struct {
+	WithinI32 int
+	BeyondI32 int
+}
+
+func main() {
+my_data := Record0{
+	WithinI32: 1705320000,
+	BeyondI32: 4085195400,
+}
+_ = my_data
+}