+++ tests/integration/cases/record_epoch_datetime_i32_overflow/Nim@v2.nim
@@ -1,0 +1,5 @@
+import json
+var my_data = %* {
+    "within_i32": {"year": 2024, "month": 1, "day": 15, "hour": 12, "minute": 0, "second": 0},
+    "beyond_i32": {"year": 2099, "month": 6, "day": 15, "hour": 8, "minute": 30, "second": 0}
+}