+++ tests/integration/cases/record_epoch_datetime_i32_overflow/Racket@v8.rkt
@@ -1,0 +1,5 @@
+#lang racket
+(define my_data (hash
+    "within_i32" "2024-01-15T12:00:00"
+    "beyond_i32" "2099-06-15T08:30:00"
+))