+++ tests/integration/cases/record_epoch_datetime_i32_overflow/Ruby@v3.rb
@@ -1,0 +1,4 @@
+my_data = {
+  "within_i32" => Time.new(2024, 1, 15, 12, 0, 0),
+  "beyond_i32" => Time.new(2099, 6, 15, 8, 30, 0),
+}