+++ tests/integration/cases/record_epoch_datetime_i32_overflow/Scheme@r7rs.scm
@@ -1,0 +1,4 @@
+(define my_data (list
+    "within_i32" "2024-01-15T12:00:00"
+    "beyond_i32" "2099-06-15T08:30:00"
+))