+++ tests/integration/cases/record_epoch_datetime_i32_overflow/VisualBasic@dotnet_6.vb
@@ -1,0 +1,7 @@
+Imports System.Collections.Generic
+Module Check
+    Dim my_data = New Dictionary(Of String, Object) From {
+        {"within_i32", "2024-01-15T12:00:00"},
+        {"beyond_i32", "2099-06-15T08:30:00"}
+    }
+End Module