+++ tests/integration/cases/record_epoch_datetime_i32_overflow/VisualBasic_combined@dotnet_6.vb
@@ -1,0 +1,16 @@
+Imports System.Collections.Generic
+Module Check
+    Sub _declaration()
+        Dim my_data = New Dictionary(Of String, Object) From {
+            {"within_i32", "2024-01-15T12:00:00"},
+            {"beyond_i32", "2099-06-15T08:30:00"}
+        }
+    End Sub
+    Sub _assignment()
+        Dim my_data As Object
+        my_data = New Dictionary(Of String, Object) From {
+            {"within_i32", "2024-01-15T12:00:00"},
+            {"beyond_i32", "2099-06-15T08:30:00"}
+        }
+    End Sub
+End Module